requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
docopt==0.6.2
exceptiongroup==1.1.3
Flask==2.3.3
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
npyscreen==4.10.5
packaging==23.1
Pillow==10.1.0
pluggy==1.3.0
pytest==7.4.2
tomli==2.0.1
Werkzeug==2.3.7
zipp==3.16.2
